Refactoring according to comments
